# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QLGD API",
        "description": "Tài liệu OpenAPI cho hệ thống Quản lý Giảng dạy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "Máy chủ API"
        }
    ],
    "paths": {
...
        "/api/lecturer/leave-requests": {
            "get": {
                "tags": [
                    "Lecturer - Nghỉ dạy"
                ],
                "summary": "Danh sách đơn xin nghỉ của giảng viên",
                "operationId": "lecturerLeaveIndex",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Lọc theo trạng thái (PENDING, APPROVED, REJECTED, CANCELED)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách phân trang",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LeaveRequestResource"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lecturer - Nghỉ dạy"
                ],
                "summary": "Tạo đơn xin nghỉ dạy",
                "operationId": "lecturerLeaveStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "reason"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "type": "integer",
                                        "example": 345
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Ốm đột xuất"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo đơn nghỉ thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LeaveRequestResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy lịch dạy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lecturer/leave-requests/{id}": {
            "get": {
                "tags": [
                    "Lecturer - Nghỉ dạy"
                ],
                "summary": "Chi tiết đơn nghỉ",
                "operationId": "lecturerLeaveShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin đơn nghỉ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LeaveRequestResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lecturer - Nghỉ dạy"
                ],
                "summary": "Hủy đơn nghỉ",
                "operationId": "lecturerLeaveDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 100
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Đã hủy"
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Chỉ hủy được đơn ở trạng thái PENDING",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Lecturer - Nghỉ dạy"
                ],
                "summary": "Chỉnh sửa đơn nghỉ (khi còn PENDING)",
                "operationId": "lecturerLeaveUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 100
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Có công tác gấp"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Đính kèm giấy xác nhận",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LeaveRequestResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Không thể chỉnh sửa đơn đã xử lý",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Thông tin xác thực không hợp lệ."
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 120
                    }
                },
                "type": "object"
            },
...
            "ScheduleItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 345
                    },
                    "session_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-21"
                    },
                    "status": {
                        "type": "string",
                        "example": "PLANNED"
                    },
                    "start_time": {
                        "type": "string",
                        "example": "07:00"
                    },
                    "end_time": {
                        "type": "string",
                        "example": "08:45"
                    },
                    "assignment": {
                        "properties": {
                            "semester_label": {
                                "type": "string",
                                "example": "2025-2026 HK1"
                            },
                            "subject": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "CT101"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Cấu trúc dữ liệu"
                                    }
                                },
                                "type": "object"
                            },
                            "classUnit": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 88
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "LHP01"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Lớp HP 01"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "room": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 12
                            },
                            "code": {
                                "type": "string",
                                "example": "B3-201"
                            },
                            "name": {
                                "type": "string",
                                "example": "Phòng B3-201"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "LeaveRequestResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 100
                    },
                    "schedule_id": {
                        "type": "integer",
                        "example": 345
                    },
                    "lecturer_id": {
                        "type": "integer",
                        "example": 7
                    },
                    "reason": {
                        "type": "string",
                        "example": "Ốm đột xuất"
                    },
                    "status": {
                        "type": "string",
                        "example": "PENDING"
                    },
                    "approved_by": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "approved_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    },
                    "schedule": {
                        "$ref": "#/components/schemas/ScheduleItem"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter the access token returned after login (omit the Bearer prefix)."
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}